Bit by bit, trying to make the codebase comprehensible
