+++ psonic/api.py
@@ -1,4 +1,8 @@
 from .psonic import *
+from .notes import *
+from .scales import *
+from .synthesizers import *
+from .effects import *
 from .samples import *
 from .samples.loops import *
 from .samples.ambient import *
@@ -10,12 +14,8 @@ from .samples.misc import *
 from .samples.percussions import *
 from .samples.tabla import *
 from .samples.vinyl import *
-
-
-
-
-
-
+from .internals.chords import *
+from .internals.scales import *
