+++ psonic/effects.py
@@ -1,0 +1,24 @@
+class FxName:
+    """
+    FX name
+    """
+
+    def __init__(self, name):
+        self.name = name
+
+## FX
+BITCRUSHER = FxName('bitcrusher')
+COMPRESSOR = FxName('compressor')
+ECHO = FxName('echo')
+FLANGER = FxName('flanger')
+KRUSH = FxName('krush')
+LPF = FxName('lpf')
+PAN = FxName('pan')
+PANSLICER = FxName('panslicer')
+REVERB = FxName('reverb')
+SLICER = FxName('slicer')
+WOBBLE = FxName('wobble')
+
+
+
+
