+++ psonic/internals/chords.py
@@ -1,0 +1,65 @@
+_CHORD_QUALITY = {
+    'major': [0, 4, 7],
+    'minor': [0, 3, 7],
+    'major7': [0, 4, 7, 11],
+    'dom7': [0, 4, 7, 10],
+    'minor7': [0, 3, 7, 10],
+    'aug': [0, 4, 8],
+    'dim': [0, 3, 6],
+    'dim7': [0, 3, 6, 9],
+    '1': [0],
+    "5": [0, 7],
+    "+5": [0, 4, 8],
+    "m+5": [0, 3, 8],
+    "sus2": [0, 2, 7],
+    "sus4": [0, 5, 7],
+    "6": [0, 4, 7, 9],
+    "m6": [0, 3, 7, 9],
+    "7sus2": [0, 2, 7, 10],
+    "7sus4": [0, 5, 7, 10],
+    "7-5": [0, 4, 6, 10],
+    "m7-5": [0, 3, 6, 10],
+    "7+5": [0, 4, 8, 10],
+    "m7+5": [0, 3, 8, 10],
+    "9": [0, 4, 7, 10, 14],
+    "m9": [0, 3, 7, 10, 14],
+    "m7+9": [0, 3, 7, 10, 14],
+    "maj9": [0, 4, 7, 11, 14],
+    "9sus4": [0, 5, 7, 10, 14],
+    "6*9": [0, 4, 7, 9, 14],
+    "m6*9": [0, 3, 9, 7, 14],
+    "7-9": [0, 4, 7, 10, 13],
+    "m7-9": [0, 3, 7, 10, 13],
+    "7-10": [0, 4, 7, 10, 15],
+    "9+5": [0, 10, 13],
+    "m9+5": [0, 10, 14],
+    "7+5-9": [0, 4, 8, 10, 13],
+    "m7+5-9": [0, 3, 8, 10, 13],
+    "11": [0, 4, 7, 10, 14, 17],
+    "m11": [0, 3, 7, 10, 14, 17],
+    "maj11": [0, 4, 7, 11, 14, 17],
+    "11+": [0, 4, 7, 10, 14, 18],
+    "m11+": [0, 3, 7, 10, 14, 18],
+    "13": [0, 4, 7, 10, 14, 17, 21],
+    "m13": [0, 3, 7, 10, 14, 17, 21],
+    "M": [0, 4, 7],
+    "m": [0, 3, 7],
+    "7": [0, 4, 7, 10],
+    "M7": [0, 4, 7, 11],
+    "m7": [0, 3, 7],
+    "augmented": [0, 4, 8],
+    "a": [0, 4, 8],
+    "diminished": [0, 3, 6],
+    "i": [0, 3, 6],
+    "diminished7": [0, 3, 6, 9],
+    "i7": [0, 3, 6, 9]
+}
+## Chord Quality (from sonic pi) ##
+MAJOR7 = "major7"
+DOM7 = "dom7"
+MINOR7 = "minor7"
+AUG = "aug"
+DIM = "dim"
+DIM7 = "dim7"
+
+
