+++ psonic/notes.py
@@ -1,0 +1,72 @@
+## Notes ##
+C2 = 36
+Cs2 = 37
+Db2 = Cs2
+D2 = 38
+Ds2 = 39
+Eb2 = Ds2
+E2 = 40
+F2 = 41
+Fs2 = 42
+Gb2 = Fs2
+G2 = 43
+Gs2 = 44
+Ab2 = Gs2
+A2 = 45
+As2 = 46
+Bb2 = As2
+B2 = 47
+C3 = 48
+Cs3 = 49
+Db3 = Cs3
+D3 = 50
+Ds3 = 51
+Eb3 = Ds3
+E3 = 52
+F3 = 53
+Fs3 = 54
+Gb3 = Fs3
+G3 = 55
+Gs3 = 56
+Ab3 = Gs3
+A3 = 57
+As3 = 58
+Bb3 = As3
+B3 = 59
+C4 = 60
+Cs4 = 61
+Db4 = Cs4
+D4 = 62
+Ds4 = 63
+Eb4 = Ds4
+E4 = 64
+F4 = 65
+Fs4 = 66
+Gb4 = Fs4
+G4 = 67
+Gs4 = 68
+Ab4 = Gs4
+A4 = 69
+As4 = 70
+Bb4 = As4
+B4 = 71
+C5 = 72
+Cs5 = 73
+Db5 = Cs5
+D5 = 74
+Ds5 = 75
+Eb5 = Ds5
+E5 = 76
+F5 = 77
+Fs5 = 78
+Gb5 = Fs5
+G5 = 79
+Gs5 = 80
+Ab5 = Gs5
+A5 = 81
+As5 = 82
+Bb5 = As5
+B5 = 83
+C6 = 84
+
+R = 0
