+++ psonic/psonic.py
@@ -1,28 +1,13 @@
-# The MIT License (MIT)
-#
-# Copyright (c) 2016 G. Völkl
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
 import random
 import time
 import threading
-from psonic.samples import Sample
+
+#TODO: Optimize imports
+from .samples import Sample
+from .synthesizers import BEEP
+from .internals.chords import _CHORD_QUALITY
+from .internals.scales import _SCALE_MODE
+
 from pythonosc import osc_message_builder  # osc support
 from pythonosc import udp_client
 
@@ -30,16 +15,6 @@ __debug = False
 
 
 ## Base Classes ##
-
-class Synth:
-    """
-    Synthesizer
-    """
-
-    def __init__(self, name):
-        self.name = name
-
-
 class ChordQuality:
     """
     Chord Quality
@@ -49,14 +24,6 @@ class ChordQuality:
         self.name = name
         self.inter = inter
 
-
-class FxName:
-    """
-    FX name
-    """
-
-    def __init__(self, name):
-        self.name = name
 
 
 class Message:
@@ -94,374 +61,14 @@ class Fx:
         pass
 
 
-## Decorator ##
-
 def in_thread(func):
+    """Thread decorator"""
     def wrapper():
         _thread = threading.Thread(target=func)
         _thread.start()
 
     return wrapper
 
-
-## Notes ##
-C2 = 36
-Cs2 = 37
-Db2 = Cs2
-D2 = 38
-Ds2 = 39
-Eb2 = Ds2
-E2 = 40
-F2 = 41
-Fs2 = 42
-Gb2 = Fs2
-G2 = 43
-Gs2 = 44
-Ab2 = Gs2
-A2 = 45
-As2 = 46
-Bb2 = As2
-B2 = 47
-C3 = 48
-Cs3 = 49
-Db3 = Cs3
-D3 = 50
-Ds3 = 51
-Eb3 = Ds3
-E3 = 52
-F3 = 53
-Fs3 = 54
-Gb3 = Fs3
-G3 = 55
-Gs3 = 56
-Ab3 = Gs3
-A3 = 57
-As3 = 58
-Bb3 = As3
-B3 = 59
-C4 = 60
-Cs4 = 61
-Db4 = Cs4
-D4 = 62
-Ds4 = 63
-Eb4 = Ds4
-E4 = 64
-F4 = 65
-Fs4 = 66
-Gb4 = Fs4
-G4 = 67
-Gs4 = 68
-Ab4 = Gs4
-A4 = 69
-As4 = 70
-Bb4 = As4
-B4 = 71
-C5 = 72
-Cs5 = 73
-Db5 = Cs5
-D5 = 74
-Ds5 = 75
-Eb5 = Ds5
-E5 = 76
-F5 = 77
-Fs5 = 78
-Gb5 = Fs5
-G5 = 79
-Gs5 = 80
-Ab5 = Gs5
-A5 = 81
-As5 = 82
-Bb5 = As5
-B5 = 83
-C6 = 84
-
-R = 0
-
-## Synthezier ##
-DULL_BELL = Synth('dull_bell')
-PRETTY_BELL = Synth('pretty_bell')
-SINE = Synth('sine')
-SQUARE = Synth('square')
-PULSE = Synth('pulse')
-SUBPULSE = Synth('subpulse')
-DTRI = Synth('dtri')
-DPULSE = Synth('dpulse')
-FM = Synth('fm')
-MOD_FM = Synth('mod_fm')
-MOD_SAW = Synth('mod_saw')
-MOD_DSAW = Synth('mod_dsaw')
-MOD_SINE = Synth('mod_sine')
-MOD_TRI = Synth('mod_tri')
-MOD_PULSE = Synth('mod_pulse')
-SUPERSAW = Synth('supersaw')
-HOOVER = Synth('hoover')
-SYNTH_VIOLIN = Synth('synth_violin')
-PLUCK = Synth('pluck')
-PIANO = Synth('piano')
-GROWL = Synth('growl')
-DARK_AMBIENCE = Synth('dark_ambience')
-DARK_SEA_HORN = Synth('dark_sea_horn')
-HOLLOW = Synth('hollow')
-ZAWA = Synth('zawa')
-NOISE = Synth('noise')
-GNOISE = Synth('gnoise')
-BNOISE = Synth('bnoise')
-CNOISE = Synth('cnoise')
-DSAW = Synth('dsaw')
-TB303 = Synth('tb303')
-BLADE = Synth('blade')
-PROPHET = Synth('prophet')
-SAW = Synth('saw')
-BEEP = Synth('beep')
-TRI = Synth('tri')
-CHIPLEAD = Synth('chiplead') # Sonic Pi 2.10
-CHIPBASS = Synth('chipbass')
-CHIPNOISE = Synth('chipnoise')
-TECHSAWS = Synth('tech_saws')  # Sonic Pi 2.11
-SOUND_IN = Synth('sound_in')
-SOUND_IN_STEREO = Synth('sound_in_stereo')
-
-## Scale Mode (from sonic pi)##
-DIATONIC = 'diatonic'
-IONIAN = 'ionian'
-MAJOR = 'major'
-DORIAN = 'dorian'
-PHRYGIAN = 'phrygian'
-LYDIAN = 'lydian'
-MIXOLYDIAN = 'mixolydian'
-AEOLIAN = 'aeolian'
-MINOR = 'minor'
-LOCRIAN = 'locrian'
-HEX_MAJOR6 = 'hex_major6'
-HEX_DORIAN = 'hex_dorian'
-HEX_PHRYGIAN = 'hex_phrygian'
-HEX_MAJOR7 = 'hex_major7'
-HEX_SUS = 'hex_sus'
-HEX_AEOLIAN = 'hex_aeolian'
-MINOR_PENTATONIC = 'minor_pentatonic'
-YU = 'yu'
-MAJOR_PENTATONIC = 'major_pentatonic'
-GONG = 'gong'
-EGYPTIAN = 'egyptian'
-SHANG = 'shang'
-JIAO = 'jiao'
-ZHI = 'zhi'
-RITUSEN = 'ritusen'
-WHOLE_TONE = 'whole_tone'
-WHOLE = 'whole'
-CHROMATIC = 'chromatic'
-HARMONIC_MINOR = 'harmonic_minor'
-MELODIC_MINOR_ASC = 'melodic_minor_asc'
-HUNGARIAN_MINOR = 'hungarian_minor'
-OCTATONIC = 'octatonic'
-MESSIAEN1 = 'messiaen1'
-MESSIAEN2 = 'messiaen2'
-MESSIAEN3 = 'messiaen3'
-MESSIAEN4 = 'messiaen4'
-MESSIAEN5 = 'messiaen5'
-MESSIAEN6 = 'messiaen6'
-MESSIAEN7 = 'messiaen7'
-SUPER_LOCRIAN = 'super_locrian'
-HIRAJOSHI = 'hirajoshi'
-KUMOI = 'kumoi'
-NEAPLOLITAN_MAJOR = 'neapolitan_major'
-BARTOK = 'bartok'
-BHAIRAV = 'bhairav'
-LOCRIAN_MAJOR = 'locrian_major'
-AHIRBHAIRAV = 'ahirbhairav'
-ENIGMATIC = 'enigmatic'
-NEAPLOLITAN_MINOR = 'neapolitan_minor'
-PELOG = 'pelog'
-AUGMENTED2 = 'augmented2'
-SCRIABIN = 'scriabin'
-HARMONIC_MAJOR = 'harmonic_major'
-MELODIC_MINOR_DESC = 'melodic_minor_desc'
-ROMANIAN_MINOR = 'romanian_minor'
-HINDU = 'hindu'
-IWATO = 'iwato'
-MELODIC_MINOR = 'melodic_minor'
-DIMISHED2 = 'diminished2'
-MARVA = 'marva'
-MELODIC_MAJOR = 'melodic_major'
-INDIAN = 'indian'
-SPANISH = 'spanish'
-PROMETHEUS = 'prometheus'
-DIMISHED = 'diminished'
-TODI = 'todi'
-LEADING_WHOLE = 'leading_whole'
-AUGMENTED = 'augmented'
-PRUVI = 'purvi'
-CHINESE = 'chinese'
-LYDIAN_MINOR = 'lydian_minor'
-I = 'i'
-II = 'ii'
-III = 'iii'
-IV = 'iv'
-V = 'v'
-VI = 'vi'
-VII = 'vii'
-VIII = 'viii'
-_ionian_sequence = [2, 2, 1, 2, 2, 2, 1]
-_hex_sequence = [2, 2, 1, 2, 2, 3]
-_pentatonic_sequence = [3, 2, 2, 3, 2]
-
-_SCALE_MODE = {
-    'diatonic': _ionian_sequence,
-    'ionian': _ionian_sequence,
-    'major': _ionian_sequence,
-    'dorian': _ionian_sequence[1:] + _ionian_sequence[:1],  # rotate 1
-    'phrygian': _ionian_sequence[2:] + _ionian_sequence[:2],  # rotate(2)
-    'lydian': _ionian_sequence[3:] + _ionian_sequence[:3],  # rotate(3)
-    'mixolydian': _ionian_sequence[4:] + _ionian_sequence[:4],  # rotate(4)
-    'aeolian': _ionian_sequence[5:] + _ionian_sequence[:5],  # rotate(5)
-    'minor': _ionian_sequence[5:] + _ionian_sequence[:5],  # rotate(5)
-    'locrian': _ionian_sequence[6:] + _ionian_sequence[:6],  # rotate(6)
-    'hex_major6': _hex_sequence,
-    'hex_dorian': _hex_sequence[1:] + _hex_sequence[:1],  # rotate(1)
-    'hex_phrygian': _hex_sequence[2:] + _hex_sequence[:2],  # rotate(2)
-    'hex_major7': _hex_sequence[3:] + _hex_sequence[:3],  # rotate(3)
-    'hex_sus': _hex_sequence[4:] + _hex_sequence[:4],  # rotate(4)
-    'hex_aeolian': _hex_sequence[5:] + _hex_sequence[:5],  # rotate(5)
-    'minor_pentatonic': _pentatonic_sequence,
-    'yu': _pentatonic_sequence,
-    'major_pentatonic': _pentatonic_sequence[1:] + _pentatonic_sequence[:1],  # rotate(1)
-    'gong': _pentatonic_sequence[1:] + _pentatonic_sequence[:1],  # rotate(1)
-    'egyptian': _pentatonic_sequence[2:] + _pentatonic_sequence[:2],  # rotate(2)
-    'shang': _pentatonic_sequence[2:] + _pentatonic_sequence[:2],  # rotate(2)
-    'jiao': _pentatonic_sequence[3:] + _pentatonic_sequence[:3],  # rotate(3)
-    'zhi': _pentatonic_sequence[4:] + _pentatonic_sequence[:4],  # rotate(4)
-    'ritusen': _pentatonic_sequence[4:] + _pentatonic_sequence[:4],  # rotate(4)
-    'whole_tone': [2, 2, 2, 2, 2, 2],
-    'whole': [2, 2, 2, 2, 2, 2],
-    'chromatic': [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
-    'harmonic_minor': [2, 1, 2, 2, 1, 3, 1],
-    'melodic_minor_asc': [2, 1, 2, 2, 2, 2, 1],
-    'hungarian_minor': [2, 1, 3, 1, 1, 3, 1],
-    'octatonic': [2, 1, 2, 1, 2, 1, 2, 1],
-    'messiaen1': [2, 2, 2, 2, 2, 2],
-    'messiaen2': [1, 2, 1, 2, 1, 2, 1, 2],
-    'messiaen3': [2, 1, 1, 2, 1, 1, 2, 1, 1],
-    'messiaen4': [1, 1, 3, 1, 1, 1, 3, 1],
-    'messiaen5': [1, 4, 1, 1, 4, 1],
-    'messiaen6': [2, 2, 1, 1, 2, 2, 1, 1],
-    'messiaen7': [1, 1, 1, 2, 1, 1, 1, 1, 2, 1],
-    'super_locrian': [1, 2, 1, 2, 2, 2, 2],
-    'hirajoshi': [2, 1, 4, 1, 4],
-    'kumoi': [2, 1, 4, 2, 3],
-    'neapolitan_major': [1, 2, 2, 2, 2, 2, 1],
-    'bartok': [2, 2, 1, 2, 1, 2, 2],
-    'bhairav': [1, 3, 1, 2, 1, 3, 1],
-    'locrian_major': [2, 2, 1, 1, 2, 2, 2],
-    'ahirbhairav': [1, 3, 1, 2, 2, 1, 2],
-    'enigmatic': [1, 3, 2, 2, 2, 1, 1],
-    'neapolitan_minor': [1, 2, 2, 2, 1, 3, 1],
-    'pelog': [1, 2, 4, 1, 4],
-    'augmented2': [1, 3, 1, 3, 1, 3],
-    'scriabin': [1, 3, 3, 2, 3],
-    'harmonic_major': [2, 2, 1, 2, 1, 3, 1],
-    'melodic_minor_desc': [2, 1, 2, 2, 1, 2, 2],
-    'romanian_minor': [2, 1, 3, 1, 2, 1, 2],
-    'hindu': [2, 2, 1, 2, 1, 2, 2],
-    'iwato': [1, 4, 1, 4, 2],
-    'melodic_minor': [2, 1, 2, 2, 2, 2, 1],
-    'diminished2': [2, 1, 2, 1, 2, 1, 2, 1],
-    'marva': [1, 3, 2, 1, 2, 2, 1],
-    'melodic_major': [2, 2, 1, 2, 1, 2, 2],
-    'indian': [4, 1, 2, 3, 2],
-    'spanish': [1, 3, 1, 2, 1, 2, 2],
-    'prometheus': [2, 2, 2, 5, 1],
-    'diminished': [1, 2, 1, 2, 1, 2, 1, 2],
-    'todi': [1, 2, 3, 1, 1, 3, 1],
-    'leading_whole': [2, 2, 2, 2, 2, 1, 1],
-    'augmented': [3, 1, 3, 1, 3, 1],
-    'purvi': [1, 3, 2, 1, 1, 3, 1],
-    'chinese': [4, 2, 1, 4, 1],
-    'lydian_minor': [2, 2, 2, 1, 1, 2, 2],
-    'i': _ionian_sequence,
-    'ii': _ionian_sequence[1:] + _ionian_sequence[:1],  # rotate(1)
-    'iii': _ionian_sequence[2:] + _ionian_sequence[:2],  # rotate(2)
-    'iv': _ionian_sequence[3:] + _ionian_sequence[:3],  # rotate(3)
-    'v': _ionian_sequence[4:] + _ionian_sequence[:4],  # rotate(4)
-    'vi': _ionian_sequence[5:] + _ionian_sequence[:5],  # rotate(5)
-    'vii': _ionian_sequence[6:] + _ionian_sequence[:6],  # rotate(6),
-    'viii': _ionian_sequence[7:] + _ionian_sequence[:7]}  # rotate(7)
-
-## Chord Quality (from sonic pi) ##
-MAJOR7 = "major7"
-DOM7 = "dom7"
-MINOR7 = "minor7"
-AUG = "aug"
-DIM = "dim"
-DIM7 = "dim7"
-
-_CHORD_QUALITY = {
-    'major': [0, 4, 7],
-    'minor': [0, 3, 7],
-    'major7': [0, 4, 7, 11],
-    'dom7': [0, 4, 7, 10],
-    'minor7': [0, 3, 7, 10],
-    'aug': [0, 4, 8],
-    'dim': [0, 3, 6],
-    'dim7': [0, 3, 6, 9],
-    '1': [0],
-    "5": [0, 7],
-    "+5": [0, 4, 8],
-    "m+5": [0, 3, 8],
-    "sus2": [0, 2, 7],
-    "sus4": [0, 5, 7],
-    "6": [0, 4, 7, 9],
-    "m6": [0, 3, 7, 9],
-    "7sus2": [0, 2, 7, 10],
-    "7sus4": [0, 5, 7, 10],
-    "7-5": [0, 4, 6, 10],
-    "m7-5": [0, 3, 6, 10],
-    "7+5": [0, 4, 8, 10],
-    "m7+5": [0, 3, 8, 10],
-    "9": [0, 4, 7, 10, 14],
-    "m9": [0, 3, 7, 10, 14],
-    "m7+9": [0, 3, 7, 10, 14],
-    "maj9": [0, 4, 7, 11, 14],
-    "9sus4": [0, 5, 7, 10, 14],
-    "6*9": [0, 4, 7, 9, 14],
-    "m6*9": [0, 3, 9, 7, 14],
-    "7-9": [0, 4, 7, 10, 13],
-    "m7-9": [0, 3, 7, 10, 13],
-    "7-10": [0, 4, 7, 10, 15],
-    "9+5": [0, 10, 13],
-    "m9+5": [0, 10, 14],
-    "7+5-9": [0, 4, 8, 10, 13],
-    "m7+5-9": [0, 3, 8, 10, 13],
-    "11": [0, 4, 7, 10, 14, 17],
-    "m11": [0, 3, 7, 10, 14, 17],
-    "maj11": [0, 4, 7, 11, 14, 17],
-    "11+": [0, 4, 7, 10, 14, 18],
-    "m11+": [0, 3, 7, 10, 14, 18],
-    "13": [0, 4, 7, 10, 14, 17, 21],
-    "m13": [0, 3, 7, 10, 14, 17, 21],
-    "M": [0, 4, 7],
-    "m": [0, 3, 7],
-    "7": [0, 4, 7, 10],
-    "M7": [0, 4, 7, 11],
-    "m7": [0, 3, 7],
-    "augmented": [0, 4, 8],
-    "a": [0, 4, 8],
-    "diminished": [0, 3, 6],
-    "i": [0, 3, 6],
-    "diminished7": [0, 3, 6, 9],
-    "i7": [0, 3, 6, 9]}
-
-## FX
-BITCRUSHER = FxName('bitcrusher')
-COMPRESSOR = FxName('compressor')
-ECHO = FxName('echo')
-FLANGER = FxName('flanger')
-KRUSH = FxName('krush')
-LPF = FxName('lpf')
-PAN = FxName('pan')
-PANSLICER = FxName('panslicer')
-REVERB = FxName('reverb')
-SLICER = FxName('slicer')
-WOBBLE = FxName('wobble')
 
 ## Module attributes ##
 _current_synth = BEEP
@@ -681,7 +288,6 @@ class Ring:
 
 
 ## Connection classes ##
-
 class SonicPi:
     """
     Communiction to Sonic Pi
@@ -775,9 +381,7 @@ class SonicPiNew:
 
 synth_server = SonicPi()
 
-
 ## system functions ##
-
 def _debug(*allargs):  # simple debug function for working in different environments
     if __debug: print(allargs)
