+++ psonic/scales.py
@@ -1,0 +1,80 @@
+## Scale Mode (from sonic pi)##
+DIATONIC = 'diatonic'
+IONIAN = 'ionian'
+MAJOR = 'major'
+DORIAN = 'dorian'
+PHRYGIAN = 'phrygian'
+LYDIAN = 'lydian'
+MIXOLYDIAN = 'mixolydian'
+AEOLIAN = 'aeolian'
+MINOR = 'minor'
+LOCRIAN = 'locrian'
+HEX_MAJOR6 = 'hex_major6'
+HEX_DORIAN = 'hex_dorian'
+HEX_PHRYGIAN = 'hex_phrygian'
+HEX_MAJOR7 = 'hex_major7'
+HEX_SUS = 'hex_sus'
+HEX_AEOLIAN = 'hex_aeolian'
+MINOR_PENTATONIC = 'minor_pentatonic'
+YU = 'yu'
+MAJOR_PENTATONIC = 'major_pentatonic'
+GONG = 'gong'
+EGYPTIAN = 'egyptian'
+SHANG = 'shang'
+JIAO = 'jiao'
+ZHI = 'zhi'
+RITUSEN = 'ritusen'
+WHOLE_TONE = 'whole_tone'
+WHOLE = 'whole'
+CHROMATIC = 'chromatic'
+HARMONIC_MINOR = 'harmonic_minor'
+MELODIC_MINOR_ASC = 'melodic_minor_asc'
+HUNGARIAN_MINOR = 'hungarian_minor'
+OCTATONIC = 'octatonic'
+MESSIAEN1 = 'messiaen1'
+MESSIAEN2 = 'messiaen2'
+MESSIAEN3 = 'messiaen3'
+MESSIAEN4 = 'messiaen4'
+MESSIAEN5 = 'messiaen5'
+MESSIAEN6 = 'messiaen6'
+MESSIAEN7 = 'messiaen7'
+SUPER_LOCRIAN = 'super_locrian'
+HIRAJOSHI = 'hirajoshi'
+KUMOI = 'kumoi'
+NEAPLOLITAN_MAJOR = 'neapolitan_major'
+BARTOK = 'bartok'
+BHAIRAV = 'bhairav'
+LOCRIAN_MAJOR = 'locrian_major'
+AHIRBHAIRAV = 'ahirbhairav'
+ENIGMATIC = 'enigmatic'
+NEAPLOLITAN_MINOR = 'neapolitan_minor'
+PELOG = 'pelog'
+AUGMENTED2 = 'augmented2'
+SCRIABIN = 'scriabin'
+HARMONIC_MAJOR = 'harmonic_major'
+MELODIC_MINOR_DESC = 'melodic_minor_desc'
+ROMANIAN_MINOR = 'romanian_minor'
+HINDU = 'hindu'
+IWATO = 'iwato'
+MELODIC_MINOR = 'melodic_minor'
+DIMISHED2 = 'diminished2'
+MARVA = 'marva'
+MELODIC_MAJOR = 'melodic_major'
+INDIAN = 'indian'
+SPANISH = 'spanish'
+PROMETHEUS = 'prometheus'
+DIMISHED = 'diminished'
+TODI = 'todi'
+LEADING_WHOLE = 'leading_whole'
+AUGMENTED = 'augmented'
+PRUVI = 'purvi'
+CHINESE = 'chinese'
+LYDIAN_MINOR = 'lydian_minor'
+I = 'i'
+II = 'ii'
+III = 'iii'
+IV = 'iv'
+V = 'v'
+VI = 'vi'
+VII = 'vii'
+VIII = 'viii'
