+++ psonic/synthesizers.py
@@ -1,0 +1,56 @@
+
+
+class Synth:
+    """
+    Synthesizer
+    """
+
+    def __init__(self, name):
+        self.name = name
+
+
+## Synthezier ##
+DULL_BELL = Synth('dull_bell')
+PRETTY_BELL = Synth('pretty_bell')
+SINE = Synth('sine')
+SQUARE = Synth('square')
+PULSE = Synth('pulse')
+SUBPULSE = Synth('subpulse')
+DTRI = Synth('dtri')
+DPULSE = Synth('dpulse')
+FM = Synth('fm')
+MOD_FM = Synth('mod_fm')
+MOD_SAW = Synth('mod_saw')
+MOD_DSAW = Synth('mod_dsaw')
+MOD_SINE = Synth('mod_sine')
+MOD_TRI = Synth('mod_tri')
+MOD_PULSE = Synth('mod_pulse')
+SUPERSAW = Synth('supersaw')
+HOOVER = Synth('hoover')
+SYNTH_VIOLIN = Synth('synth_violin')
+PLUCK = Synth('pluck')
+PIANO = Synth('piano')
+GROWL = Synth('growl')
+DARK_AMBIENCE = Synth('dark_ambience')
+DARK_SEA_HORN = Synth('dark_sea_horn')
+HOLLOW = Synth('hollow')
+ZAWA = Synth('zawa')
+NOISE = Synth('noise')
+GNOISE = Synth('gnoise')
+BNOISE = Synth('bnoise')
+CNOISE = Synth('cnoise')
+DSAW = Synth('dsaw')
+TB303 = Synth('tb303')
+BLADE = Synth('blade')
+PROPHET = Synth('prophet')
+SAW = Synth('saw')
+BEEP = Synth('beep')
+TRI = Synth('tri')
+CHIPLEAD = Synth('chiplead') # Sonic Pi 2.10
+CHIPBASS = Synth('chipbass')
+CHIPNOISE = Synth('chipnoise')
+TECHSAWS = Synth('tech_saws')  # Sonic Pi 2.11
+SOUND_IN = Synth('sound_in')
+SOUND_IN_STEREO = Synth('sound_in_stereo')
+
+
